requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
whitenoise==6.6.0
gunicorn==22.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9